🚚 engine.h: Extract `Point`, `Direction`, `clamp`
`engine.h` is getting quite bloated. Moves this code to their own files.
