+++ Source/.clang-tidy
@@ -81,3 +81,5 @@ CheckOptions:
 
   # `int8_t` aren't used as chars, disable misleading warning.
   - { key: bugprone-signed-char-misuse.CharTypdefsToIgnore, value: "std::int8_t" }
+
+  - { key: misc-non-private-member-variables-in-classes.IgnoreClassesWithAllMemberVariablesBeingPublic, value: true }
