+++ Source/dead.h
@@ -8,6 +8,7 @@
 #include <array>
 #include <cstdint>
 
+#include "engine/point.hpp"
 #include "engine.h"
 
 namespace devilution {
