+++ Source/engine/direction.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <cstdint>
+
+namespace devilution {
+
+enum Direction : std::uint8_t {
+	DIR_S,
+	DIR_SW,
+	DIR_W,
+	DIR_NW,
+	DIR_N,
+	DIR_NE,
+	DIR_E,
+	DIR_SE,
+	DIR_OMNI,
+};
+
+} // namespace devilution
