+++ Source/items.h
@@ -9,6 +9,7 @@
 
 #include "DiabloUI/ui_item.h"
 #include "engine/animationinfo.h"
+#include "engine/point.hpp"
 #include "engine.h"
 #include "itemdat.h"
 #include "utils/stdcompat/optional.hpp"
