+++ Source/loadsave.cpp
@@ -15,6 +15,7 @@
 #include "dead.h"
 #include "doom.h"
 #include "engine.h"
+#include "engine/point.hpp"
 #include "init.h"
 #include "inv.h"
 #include "lighting.h"
