+++ Source/multi.cpp
@@ -12,6 +12,7 @@
 #include "DiabloUI/diabloui.h"
 #include "diablo.h"
 #include "dthread.h"
+#include "engine/point.hpp"
 #include "mainmenu.h"
 #include "nthread.h"
 #include "options.h"
