+++ Source/qol/itemlabels.h
@@ -5,6 +5,8 @@
 */
 #pragma once
 
+#include "engine.h"
+
 namespace devilution {
 
 void ToggleItemLabelHighlight();
