+++ Source/utils/stdcompat/cstddef.hpp
@@ -1,15 +1,14 @@
 #pragma once
 
-#include <cstddef>
-#ifdef __has_include
+#include <cstddef> // IWYU pragma: export
+
 #if defined(__cplusplus) && __cplusplus >= 201703L
 namespace devilution {
 using byte = std::byte;
-}
+} // namespace devilution
 #else
 #include <cstdint>
 namespace devilution {
 using byte = std::uint8_t;
-}
-#endif
+} // namespace devilution
 #endif
